refactor(focus-controller): extract to correctly package
