+++ app/src/main/java/com/neo/speaktouch/controller/FocusController.kt
@@ -16,20 +16,17 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.neo.speaktouch.intercepter
+package com.neo.speaktouch.controller
 
 import android.view.accessibility.AccessibilityNodeInfo
-import androidx.core.view.accessibility.AccessibilityNodeInfoCompat
 import com.neo.speaktouch.model.NodeFilter
 import com.neo.speaktouch.utils.extension.Direction
 import com.neo.speaktouch.utils.extension.ancestors
 import com.neo.speaktouch.utils.extension.descendants
 import com.neo.speaktouch.utils.extension.getFocusedOrNull
-import com.neo.speaktouch.utils.extension.getNearestAncestor
 import com.neo.speaktouch.utils.extension.indexOfChild
 import com.neo.speaktouch.utils.extension.lastIndex
 import com.neo.speaktouch.utils.extension.performFocus
-import com.neo.speaktouch.utils.`object`.NodeValidator
 
 class FocusController(
     private val a11yNodeInfoRoot: () -> AccessibilityNodeInfo
